chore: setup env and refactor api
