+++ client/.env.example
@@ -1,0 +1,1 @@
+VITE_API_BASE_URL=http://localhost:3000/api
