+++ client/src/api/apiClient.ts
@@ -1,0 +1,7 @@
+import axios from 'axios';
+
+const apiClient = axios.create({
+  baseURL: import.meta.env.VITE_API_BASE_URL,
+});
+
+export default apiClient;
